requirements.txt
fastapi==0.119.1
uvicorn[standard]==0.38.0
Pillow==12.0.0
watchdog==6.0.0
rawpy==0.25.1
authlib==1.2.1
httpx==0.25.2
itsdangerous==2.1.2
python-multipart==0.0.6
